requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
python-telegram-bot==12.1.1
requests==2.22.0
six==1.12.0
spotipy==2.4.4
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.5
wrapt==1.11.2
